pyproject.toml
[project]
name = "nexus-platform"
version = "1.0.0"
description = "Production-ready NEXUS platform with integrated analytics module"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    # Core Framework
    "streamlit>=1.31.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",

    # AI Integration
    "anthropic>=0.18.0",
    "openai>=1.12.0",

    # Database
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",

    # Caching & Task Queue
    "redis>=5.0.1",
    "celery>=5.3.6",
    "flower>=2.0.1",

    # Data Processing
    "pandas>=2.2.0",
    "numpy>=1.26.3",
    "polars>=0.20.6",
    "pyarrow>=15.0.0",

    # Time Series
    "influxdb-client>=1.40.0",

    # Machine Learning
    "scikit-learn>=1.4.0",
    "prophet>=1.1.5",
    "statsmodels>=0.14.1",

    # Visualization
    "plotly>=5.18.0",
    "altair>=5.2.0",
    "matplotlib>=3.8.2",
    "seaborn>=0.13.2",

    # Data Validation
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",

    # Utilities
    "python-dotenv>=1.0.1",
    "pytz>=2024.1",
    "python-dateutil>=2.8.2",
    "click>=8.1.7",

    # Export Formats
    "openpyxl>=3.1.2",
    "reportlab>=4.0.9",
    "fpdf2>=2.7.8",

    # HTTP & API
    "httpx>=0.26.0",
    "aiohttp>=3.9.3",

    # Monitoring
    "prometheus-client>=0.19.0",

    # Serialization
    "orjson>=3.9.13",
    "msgpack>=1.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.4",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "ruff>=0.2.1",
    "mypy>=1.8.0",
    "black>=24.1.1",
    "isort>=5.13.2",
    "pre-commit>=3.6.0",
    "ipython>=8.21.0",
    "ipdb>=0.13.13",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["modules", "shared"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = """
    -v
    --tb=short
    --cov=modules
    --cov=shared
    --cov-report=html
    --cov-report=term-missing:skip-covered
    --cov-fail-under=80
    --asyncio-mode=auto
"""
asyncio_mode = "auto"

[tool.coverage.run]
source = ["modules", "shared"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
